+++ packages/smarthr-ui/src/components/FormControl/stories/FormControl.stories.tsx
@@ -10,6 +10,7 @@ import { StatusLabel } from '../../StatusLabel'
 import { STYLE_TYPE_MAP } from '../../Text'
 import { Textarea } from '../../Textarea'
 import { FormControl } from '../FormControl'
+import { FaAddressBookIcon } from '../../Icon'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
@@ -50,8 +51,7 @@ export default {
     supplementaryMessage: { control: 'text' },
   },
   args: {
-    title: 'フォームコントロール',
-
+    label: 'フォームコントロール',
     children: <Input />,
   },
   parameters: {
@@ -72,27 +72,73 @@ export const Playground: StoryObj<typeof FormControl> = {
   args: {},
 }
 
-export const Title: StoryObj<typeof FormControl> = {
-  name: 'title',
+export const Label: StoryObj<typeof FormControl> = {
+  name: 'label',
   args: {
-    title: '入力要素に紐づく名前',
+    label: '入力要素に紐づく名前',
   },
 }
 
-export const TitleType: StoryObj<typeof FormControl> = {
-  name: 'titleType',
+export const LabelStyleType: StoryObj<typeof FormControl> = {
+  name: 'label.styleType',
   render: (args) => (
     <Stack>
-      {[undefined, ...Object.keys(STYLE_TYPE_MAP)].map((titleType) => (
+      {[undefined, ...Object.keys(STYLE_TYPE_MAP)].map((labelType) => (
         <FormControl
           {...args}
-          title={titleType ?? 'undefined'}
-          titleType={titleType as any}
-          key={titleType}
+          label={
+            labelType
+              ? {
+                  text: labelType,
+                  styleType: labelType,
+                }
+              : 'undefined'
+          }
+          key={labelType}
         />
       ))}
     </Stack>
   ),
+}
+
+export const DangerouslyTitleHidden: StoryObj<typeof FormControl> = {
+  name: 'label.dangerouslyHide（利用注意）',
+  args: {
+    label: {
+      text: 'フォームコントロール',
+      dangerouslyHide: true,
+    },
+  },
+}
+
+export const HtmlFor: StoryObj<typeof FormControl> = {
+  name: 'label.htmlFor',
+  args: {
+    label: {
+      text: 'フォームコントロール',
+      htmlFor: 'input-id',
+    },
+  },
+}
+
+export const LabelId: StoryObj<typeof FormControl> = {
+  name: 'label.id',
+  args: {
+    label: {
+      text: 'フォームコントロール',
+      id: 'label-id',
+    },
+  },
+}
+
+export const LabelIcon: StoryObj<typeof FormControl> = {
+  name: 'label.icon',
+  args: {
+    label: {
+      text: 'フォームコントロール',
+      icon: <FaAddressBookIcon />,
+    },
+  },
 }
 
 export const SubActionArea: StoryObj<typeof FormControl> = {
@@ -104,27 +150,6 @@ export const SubActionArea: StoryObj<typeof FormControl> = {
         <div>サブアクションエリア（end)</div>
       </Cluster>
     ),
-  },
-}
-
-export const DangerouslyTitleHidden: StoryObj<typeof FormControl> = {
-  name: 'dangerouslyTitleHidden（利用注意）',
-  args: {
-    dangerouslyTitleHidden: true,
-  },
-}
-
-export const HtmlFor: StoryObj<typeof FormControl> = {
-  name: 'htmlFor',
-  args: {
-    htmlFor: 'input-id',
-  },
-}
-
-export const LabelId: StoryObj<typeof FormControl> = {
-  name: 'labelId',
-  args: {
-    labelId: 'label-id',
   },
 }