+++ packages/smarthr-ui/src/components/RangeSeparator/index.ts
@@ -1,1 +1,0 @@
-export { RangeSeparator } from './RangeSeparator'