+++ packages/smarthr-ui/src/components/ResponseMessage/stories/ResponseMessage.stories.tsx
@@ -12,13 +12,12 @@ export default {
       control: 'select',
       options: Object.keys(classNameGenerator.variants.type),
     },
-    size: { name: 'size（非推奨）' },
+    size: { name: 'size' },
     color: { table: { disable: true } },
   },
   args: {
     children: 'レスポンスメッセージ',
     type: 'info',
-    size: 'M',
     alt: '',
     right: false,
     iconGap: 0.25,
@@ -42,16 +41,14 @@ export const Type: StoryObj<typeof ResponseMessage> = {
 }
 
 export const Size: StoryObj<typeof ResponseMessage> = {
-  name: 'size（非推奨）',
+  name: 'size',
   render: (args) => (
     <Stack align="flex-start">
-      <ResponseMessage {...args} size="XXS" />
-      <ResponseMessage {...args} size="XS" />
-      <ResponseMessage {...args} size="S" />
-      <ResponseMessage {...args} size="M" />
-      <ResponseMessage {...args} size="L" />
-      <ResponseMessage {...args} size="XL" />
-      <ResponseMessage {...args} size="XXL" />
+      {([undefined, 'M', 'S', 'XS'] as const).map((size) => (
+        <ResponseMessage {...args} size={size}>
+          {size || 'size未指定はMと同サイズ'}
+        </ResponseMessage>
+      ))}
     </Stack>
   ),
 }