+++ sandbox/next/e2e/rsc.test.ts
@@ -29,8 +29,6 @@ const SERVER_COMPONENTS = [
   'MonthPicker',
   'NotificationBar',
   'PageCounter',
-  'Pagination',
-  'RangeSeparator',
   'ResponseMessage',
   'SearchInput',
   'SideMenu',
@@ -114,23 +112,25 @@ const CLIENT_COMPONENTS: string[] = [
   'Textarea',       // オートフォーカスや文字数カウントのための状態管理
   'ThCheckbox',     // 多言語化対応のため
   'Tooltip',        // 開閉状態の管理
+  'DateFormatter',  // 多言語化対応のため
 ]
 
 /**
  * サーバーコンポーネント内では利用できないコンポーネント一覧(アルファベット順)
  */
 const DISABLED_COMPONENTS = [
-  'ActionDialog',            // 必須のイベントハンドラをもつため
-  'Calendar',                // 必須のイベントハンドラをもつため
-  'DropZone',                // 必須のイベントハンドラをもつため
-  'FilterDropdown',          // 必須のイベントハンドラをもつため
-  'FormDialog',              // 必須のイベントハンドラをもつため
-  'MessageDialog',           // 必須のイベントハンドラをもつため
-  'MessageDialogContent',    // 必須のイベントハンドラを持つため
+  'ActionDialog',               // 必須のイベントハンドラをもつため
+  'Calendar',                   // 必須のイベントハンドラをもつため
+  'DropZone',                   // 必須のイベントハンドラをもつため
+  'FilterDropdown',             // 必須のイベントハンドラをもつため
+  'FormDialog',                 // 必須のイベントハンドラをもつため
+  'MessageDialog',              // 必須のイベントハンドラをもつため
+  'MessageDialogContent',       // 必須のイベントハンドラを持つため
   'RemoteTriggerActionDialog',  // 必須のイベントハンドラをもつため
   'RemoteTriggerFormDialog',    // 必須のイベントハンドラをもつため
-  'SortDropdown',           // 必須のイベントハンドラをもつため
-  'TabItem',                // 必須のイベントハンドラをもつため
+  'SortDropdown',               // 必須のイベントハンドラをもつため
+  'TabItem',                    // 必須のイベントハンドラをもつため
+  'Pagination',                 // 必須のイベントハンドラをもつため
 ]
 
 test.describe('RSC対応コンポーネントがRSCで利用できること', () => {