+++ sandbox/next/src/app/rsc_test/DateFormatter/page.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { DateFormatter } from 'smarthr-ui'
+
+import { RSCChecker } from '../components/RSCChecker'
+export default function DateFormatterPage() {
+  return (
+    <>
+      <RSCChecker actualComponent={DateFormatter} />
+      <DateFormatter date={new Date(2024, 10, 15)} />
+    </>
+  )
+}